Added Motor feedback. Fixed UARTs
Added motor feedback interrupt reads with appropriate timers. The UARTT
transmit interrupt type is changed in the new harmony config file (use
transmit FIFO is empty interrupt type). Also, transmit now only fills
the 6-byte deep buffer and then exits the interrupt until the transmit
buffer is full. It used to try to push all of the queued bytes into the
buffer which would lead to blocking. this was fixed by updating the
while() loop used to populate the transmit buffer with an incremental
integer that ensures we put in 6 bytes at a type whenever the transmit
FIFO buffer is empty.